- refreshed fields after session is booted. because some fields may be changed for example for some users etc... - fields removeFromForm/invisible/disabled are fully skipped from request now and wont be passed into db. (security feature)

+++ src/Controllers/Crud/Concerns/CRUDRelations.php
@@ -13,6 +13,12 @@ trait CRUDRelations
     {
         foreach ($model->getFields() as $key => $field) {
             if (array_key_exists('belongsToMany', $field)) {
+                //If field was deleted from request
+                //We does not want update this value.
+                if ( $request->isRemovedFieldFromRequest($key) === true ){
+                    continue;
+                }
+
                 $properties = $model->getRelationProperty($key, 'belongsToMany');
 
                 DB::table($properties[3])->where($properties[6], $row->getKey())->delete();
